// YOUCOM.ReserVook/src/app/features/web/master/planconvert/model/planconvert.model.ts
export class PlanConvertInfo {
  companyNo: string;
  scCd: string;

  scPackagePlanCd: string;
  scMealCond: string;
  scSpecMealCond: string;
  itemCode: string;

  updateCnt: number;
  programId: string;
  createClerk: string;
  createMachineNo: string;
  createMachine: string;
  createDatetime: string;
  updateClerk: string;
  updateMachineNo: string;
  updateMachine: string;
  updateDatetime: string;
  status: string;

  cdName: string;
  itemName: string;
}